ci: predictable server builds and port option
Implement the ability to create predictable server builds so that the
workflows can validate that it's been updated as needed.

Add the workflow lint which validates that everything has been built
and lint fixes applied.

Switch running the server to use -jar to reduce complexity now the
manifest is correctly created.

Add missing cli jars not needed for source runs, causing them to be
missed in previous commit.

Move scripts into their own folder to keep the root directory cleaner.

Add lint:fix and build changes, unrelated to this as they hadn't been
run in the past. Also renamed the server npm jobs so they are consistent
and shorter.

Removed try catch in server runCodeNarc as it was resulting in
exceptions being logged twice.

Major rework of CodeNarcServer to address race conditions in particular
hangs on shutdown which was causing constant random test failures.

Switch logging to logback which is more configurable including being
able to write to log files.

Move the server tests into their own file as they have been the flaky
tests.